refactor(storage): rename DeltaWriter to SilverWriter for Medallion naming consistency
Unifies Writer class naming with Medallion architecture layer convention:
- BronzeWriter - writes to Bronze layer (JSONL + zstd)
- SilverWriter - writes to Silver layer (Delta Lake) [NEW NAME]
- GoldWriter - writes to Gold layer (Delta Lake with SCD Type 2)

Changes:
- Create new SilverWriter class with identical functionality
- Convert DeltaWriter to deprecated subclass with DeprecationWarning
- Update composition layer to use SilverWriter directly
- Update documentation and tests to reference SilverWriter
- Maintain backward compatibility: DeltaWriter alias still works

The deprecation follows RULES.md §6.2 (14-day deprecation period).
