+++ src/bioetl/composition/factories/storage.py
@@ -18,8 +18,11 @@ from __future__ import annotations
 
 # Re-export writers for test patching compatibility
 from bioetl.infrastructure.storage.bronze_writer import BronzeWriter
+
+# DeltaWriter is deprecated, use SilverWriter instead
 from bioetl.infrastructure.storage.delta_writer import DeltaWriter
 from bioetl.infrastructure.storage.gold_writer import GoldWriter
+from bioetl.infrastructure.storage.silver_writer import SilverWriter
 
 # Re-export from split modules for backward compatibility
 from .storage_adapter import StorageAdapter
@@ -27,8 +30,9 @@ from .storage_factory import StorageContext, StorageFactory
 
 __all__ = [
     "BronzeWriter",
-    "DeltaWriter",
+    "DeltaWriter",  # Deprecated: use SilverWriter instead
     "GoldWriter",
+    "SilverWriter",
     "StorageAdapter",
     "StorageContext",
     "StorageFactory",
