+++ tests/unit/infrastructure/storage/test_delta_writer.py
@@ -611,7 +611,7 @@ class TestDeltaWriterErrorHandling:
         )
 
         with patch(
-            "bioetl.infrastructure.storage.delta_writer.DeltaTable",
+            "bioetl.infrastructure.storage.silver_writer.DeltaTable",
             delta_table_mock,
         ):
             writer = DeltaWriter(
@@ -666,7 +666,7 @@ class TestDeltaWriterErrorHandling:
         )
 
         with patch(
-            "bioetl.infrastructure.storage.delta_writer.DeltaTable",
+            "bioetl.infrastructure.storage.silver_writer.DeltaTable",
             delta_table_mock,
         ):
             writer = DeltaWriter(
@@ -694,7 +694,7 @@ class TestDeltaWriterSchemaDrift:
         from bioetl.infrastructure.storage.delta_writer import DeltaWriter
 
         with patch(
-            "bioetl.infrastructure.storage.delta_writer.DeltaTable",
+            "bioetl.infrastructure.storage.silver_writer.DeltaTable",
             side_effect=DeltaTableNotFoundError("Not found"),
         ):
             writer = DeltaWriter(
@@ -720,7 +720,7 @@ class TestDeltaWriterSchemaDrift:
         mock_table.schema.return_value = mock_delta_schema
 
         with patch(
-            "bioetl.infrastructure.storage.delta_writer.DeltaTable",
+            "bioetl.infrastructure.storage.silver_writer.DeltaTable",
             return_value=mock_table,
         ):
             writer = DeltaWriter(
@@ -748,7 +748,7 @@ class TestDeltaWriterSchemaDrift:
         mock_table.schema.return_value = mock_delta_schema
 
         with patch(
-            "bioetl.infrastructure.storage.delta_writer.DeltaTable",
+            "bioetl.infrastructure.storage.silver_writer.DeltaTable",
             return_value=mock_table,
         ):
             writer = DeltaWriter(
@@ -797,7 +797,7 @@ class TestDeltaWriterSchemaDrift:
         ]
 
         with patch(
-            "bioetl.infrastructure.storage.delta_writer.DeltaTable",
+            "bioetl.infrastructure.storage.silver_writer.DeltaTable",
             return_value=mock_table,
         ):
             writer = DeltaWriter(
@@ -827,7 +827,7 @@ class TestDeltaWriterSchemaDrift:
         mock_table.schema.return_value = mock_delta_schema
 
         with patch(
-            "bioetl.infrastructure.storage.delta_writer.DeltaTable",
+            "bioetl.infrastructure.storage.silver_writer.DeltaTable",
             return_value=mock_table,
         ):
             writer = DeltaWriter(
@@ -854,7 +854,7 @@ class TestDeltaWriterSchemaDrift:
         mock_table.schema.return_value = mock_delta_schema
 
         with patch(
-            "bioetl.infrastructure.storage.delta_writer.DeltaTable",
+            "bioetl.infrastructure.storage.silver_writer.DeltaTable",
             return_value=mock_table,
         ):
             writer = DeltaWriter(
@@ -891,7 +891,7 @@ class TestDeltaWriterSchemaDrift:
         mock_table.schema.return_value = mock_delta_schema
 
         with patch(
-            "bioetl.infrastructure.storage.delta_writer.DeltaTable",
+            "bioetl.infrastructure.storage.silver_writer.DeltaTable",
             return_value=mock_table,
         ):
             writer = DeltaWriter(
@@ -909,7 +909,7 @@ class TestDeltaWriterSchemaDrift:
         from bioetl.infrastructure.storage.delta_writer import DeltaWriter
 
         with patch(
-            "bioetl.infrastructure.storage.delta_writer.DeltaTable",
+            "bioetl.infrastructure.storage.silver_writer.DeltaTable",
             side_effect=DeltaTableNotFoundError("Not found"),
         ):
             writer = DeltaWriter(
@@ -934,7 +934,7 @@ class TestDeltaWriterSchemaDrift:
         mock_table.schema.return_value = mock_delta_schema
 
         with patch(
-            "bioetl.infrastructure.storage.delta_writer.DeltaTable",
+            "bioetl.infrastructure.storage.silver_writer.DeltaTable",
             return_value=mock_table,
         ):
             writer = DeltaWriter(
@@ -986,7 +986,7 @@ class TestDeltaWriterSchemaDrift:
         mock_table.schema.return_value = mock_delta_schema
 
         with patch(
-            "bioetl.infrastructure.storage.delta_writer.DeltaTable",
+            "bioetl.infrastructure.storage.silver_writer.DeltaTable",
             return_value=mock_table,
         ):
             writer = DeltaWriter(
@@ -1245,11 +1245,11 @@ class TestDeltaWriterWriteModePolicy:
 
         with (
             patch(
-                "bioetl.infrastructure.storage.delta_writer.DeltaTable",
+                "bioetl.infrastructure.storage.silver_writer.DeltaTable",
                 side_effect=DeltaTableNotFoundError("Not found"),
             ),
             patch(
-                "bioetl.infrastructure.storage.delta_writer.write_deltalake"
+                "bioetl.infrastructure.storage.silver_writer.write_deltalake"
             ) as mock_write,
         ):
             writer = DeltaWriter(
@@ -1291,11 +1291,11 @@ class TestDeltaWriterWriteModePolicy:
 
         with (
             patch(
-                "bioetl.infrastructure.storage.delta_writer.DeltaTable",
+                "bioetl.infrastructure.storage.silver_writer.DeltaTable",
                 side_effect=DeltaTableNotFoundError("Not found"),
             ),
             patch(
-                "bioetl.infrastructure.storage.delta_writer.write_deltalake"
+                "bioetl.infrastructure.storage.silver_writer.write_deltalake"
             ) as mock_write,
         ):
             writer = DeltaWriter(
